% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{acum}
\alias{acum}
\title{Acumula series em percentual ao mes em m meses}
\usage{
acum(x, m = 12)
}
\arguments{
\item{x}{A time series univariate}

\item{m}{number of monthes}
}
\value{
A time series univariate
}
\description{
Transforma uma serie mensal dada em percentual ao mes 
em uma serie mensal com percentual nos ultimos m meses
}
