% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{adf.lag}
\alias{adf.lag}
\title{Seleciona a defasagem para o teste ADF}
\usage{
adf.lag(y, type = c("none", "drift", "trend"), lag.max = 15,
  selectlags = c("AIC", "BIC"))
}
\arguments{
\item{y}{The vector tested for a unit root}

\item{type}{Test type, either "none", "drift" or "trend".}

\item{lag.max}{The maximum number of lags considered}

\item{selectlags}{Lag selection can be achieved according to the Akaike "AIC" or the Bayes "BIC" information criteria.}
}
\value{
A defasagem selecionda
}
\description{
Seleciona a defasagem para o teste ADF segundo o criterio
AIC ou BIC para um maximo de defasagens
}
