% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovecm.R
\name{auto.VECM}
\alias{auto.VECM}
\title{Fit best VECM model to multivariate time series}
\usage{
auto.VECM(y, max.p = 6, ic = c("SC", "HQ", "AIC", "FPE"), beta = NULL,
  include = c("const", "trend", "none", "both"), LRinclude = c("none",
  "const", "trend", "both"), ...)
}
\arguments{
\item{y}{A multivariate time series}

\item{max.p}{Determines the highest lag order for lag length selection 
according to the choosen ic.}

\item{ic}{Information criterion to be used in model selection.}

\item{beta}{Imposed cointegrating value.}

\item{include}{Type of deterministic regressors to include}

\item{LRinclude}{Type of deterministic regressors to include in the 
long-term relationship. Can also be a matrix with exogeneous 
regressors (2OLS only)}

\item{...}{Other arguments for \link[tsDyn]{VECM}.}
}
\value{
A list with class attribute 'autovecm' holding the following elements: 

  \item{\code{fit}}{modelo estimado da class \code{VECM}.}
  \item{\code{d}}{ordem de diferenciacao das series.}
  \item{\code{y}}{series das variaveis originais.}
  \item{\code{ic}}{criterios de informacao do modelo selecionado.}
}
\description{
Returns best VECM model according to either SC, HQ, AIC or FPE value.
The function conducts a search over possible model within the order 
constraints provided.
}

