% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoceemdan.R
\name{autoceemdan}
\alias{autoceemdan}
\title{Auto CEEMDAN decomposition}
\usage{
autoceemdan(x)
}
\arguments{
\item{x}{Vector of length n. The time serie to decompose}
}
\value{
Time series object of class "mts" where 
series corresponds to IMFs of the input time serie, 
with the last series being the final residual
}
\description{
Wrapper for \link[Rlibeemd]{ceemdan}. Apply the \link[Rlibeemd]{ceemdan}
function for noise_strength in range 0.01 at 0.99 and return the means
}
\references{
M. Torres et al, "A Complete Ensemble 
Empirical Mode Decomposition with Adaptive Noise" 
IEEE Int. Conf. on Acoust., Speech and Signal Proc. 
ICASSP-11, (2011) 4144-4147

N. E. Huang, Z. Shen and S. R. Long, "A new view of
nonlinear water waves: The Hilbert spectrum", Annual
 Review of Fluid Mechanics, Vol. 31 (1999) 417-457
}

