% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{comp.test}
\alias{comp.test}
\title{Test for Forecast Encompassing}
\usage{
comp.test(y, fA, fB, h)
}
\arguments{
\item{y}{The observed time series.}

\item{fA}{Forecast from method A.}

\item{fB}{Forecast from method B.}

\item{h}{Length ahead of the forecast.}
}
\value{
An object of class "coeftest" which is 
essentially a coefficient matrix with columns 
containing the estimates, associated standard 
errors, test statistics and p values.
}
\description{
The Test for Forecast Encompassing compares the forecast of two forecast methods.
}
\details{
The null hypothesis is that the method A forecast encompasses
of method B, i.e., method B have the same information
of method A. The alternative hypothesis is that method B 
have additional information.
}
