% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovecm.R
\name{ect}
\alias{ect}
\title{Estimation of error correction term}
\usage{
ect(y, beta = NULL, LRinclude = c("none", "const", "trend", "both"))
}
\arguments{
\item{y}{multivariate time series}

\item{beta}{imposed cointegrating value. If null, values will be estimated}

\item{LRinclude}{Type of deterministic regressors to include in the long-term
relationship. Can also be a matrix with exogeneous regressors}
}
\value{
error correction term matrix \eqn{ECT= \beta'X_{t-1}}
}
\description{
Estimate error correction term (ect) by OLS
}
\details{
The arg beta allows to specify constrained cointegrating values, 
leading to \eqn{ECT= \beta'X_{t-1}}. It should be specified as a \eqn{K \times r} 
matrix. In case of \eqn{r=1}, can also be specified as a vector. Note that
the vector should be normalised, with the first value to 1, and
the next values showing the opposite sign in the long-run relationship \eqn{-\beta}. 
In case the vector has \eqn{K-1} values, setting \eqn{(1, -\beta)}. Note finally 
one should provide values for all the coefficients (eventually except for special
case of r=1 and k-1).
}

