% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovecm.R
\name{f2list}
\alias{f2list}
\title{Transform the forecast of an mforecast object for a list}
\usage{
f2list(x)
}
\arguments{
\item{x}{An object of class "mforecast".}
}
\value{
A list of matrices containing the forecasted values with
lower and upper bounds
}
\description{
Transforma as previsoes forneccida por forecast.autovar para uma
lista contendo as previsoes e o intervalo de previsao
}

