% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovecm.R
\name{forecast.autovecm}
\alias{forecast.autovecm}
\title{Forecasting time series com VECM model}
\usage{
\method{forecast}{autovecm}(object, h = 10, level = c(80, 95),
  fan = FALSE, ...)
}
\arguments{
\item{object}{An object of class "autovecm".}

\item{h}{Number of periods for forecasting.}

\item{level}{Confidence level for prediction intervals.}

\item{fan}{If TRUE, level is set to seq(51,99,by=3). This is 
suitable for fan plots.}

\item{...}{Other arguments.}
}
\value{
An object of class "mforecast". An object of class 
"mforecast" is a list containing at least the following elements:

\item{model}{A list containing information about the fitted 
model}
\item{method}{The name of the forecasting method as a character
 string}
\item{mean}{Point forecasts as a time series}
\item{lower}{Lower limits for prediction intervals}
\item{upper}{Upper limits for prediction intervals}
\item{level}{The confidence values associated with the 
prediction intervals}
\item{x}{The original time series (either \code{object} itself
 or the time series used to create the model stored as \code{object}).}
\item{residuals}{Residuals from the fitted model. 
That is x minus fitted values.}
\item{fitted}{Fitted values (one-step forecasts)}
}
\description{
Returns forecasts and other information for VECM models
}

