% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{glag}
\alias{glag}
\title{Generalized Lag}
\usage{
glag(x, k = -1)
}
\arguments{
\item{x}{A vector or univariate time series}

\item{k}{The number of lags}
}
\value{
A vector object with the same length of x.
}
\description{
Compute a lagged version of a vector like time series
}
