% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{marques}
\alias{marques}
\title{Testa os criterios de Marques et al (2003)}
\usage{
marques(x, y)
}
\arguments{
\item{x}{serie do nucleo}

\item{y}{serie da inflacao}
}
\value{
um vetor com os teste "ADF", "t alpha", "t gamma", 't lambda', 'F thetas'
}
\description{
Testa os criterios de Marques et al (2003) que uma 
medida de nucleo da inflacao deve atender
}
