% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{mdd}
\alias{mdd}
\title{Estima e Prever com modelo de defadsagem distribuida (mdd)}
\usage{
mdd(yh, yt, x, h, m, p)
}
\arguments{
\item{yh}{(ts) serie a ser prevista \eqn{y_{t+h}^h}}

\item{yt}{(ts) preditor \eqn{y_{t-j+1}}}

\item{x}{(ts) preditor}

\item{h}{horizonte de previsao}

\item{m}{numero de defasagens de x}

\item{p}{numero de defasagens de yt}
}
\value{
uma lista com: fit(dyn): o modelo estimado e 
fcast(ts): valor previsto para \eqn{yh_{t+h}}
}
\description{
Estima uma regressao de yh sobre x e seus m lags e 
sobre yt e seus p lags
}
\details{
yh igual a yt ou a alguma transformacao de yt, esse modelo
de previsao se baseia em Stock e Watson 1998
}
