% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplots.R
\name{n2t}
\alias{n2t}
\title{#' Teste}
\usage{
n2t(x, dig = 3, decimal.mark = ",")
}
\arguments{
\item{x}{object to be coerced to formatted text}

\item{dig}{the minimum number of digits to the right 
of the decimal point in formatting real numbers 
in non-scientific formats}

\item{decimal.mark}{set the ‘decimal mark’ other than "."}
}
\value{
An object of similar structure to x containing character 
representations of the elements of the first argument x in 
formatted text
}
\description{
lmplot <- function(x, y, x.name = NULL, y.name = NULL, title = NULL){
  if(length(x)!=length(y))
    stop("x and y are different length")
  if(is.null(x.name)){
    if(is.null(names(x))){
      x.name <- "x"
    }else{
      x.name <- names(x)
    }
  }
  if(is.null(y.name)){
    if(is.null(names(y))){
      y.name <- "y"
    }else{
      y.name <- names(y)
    }
  }
  if(is.null(title)){
    title <- paste(x.name, " vs", y.name)
  }
  dat <- data.frame(x, y, serie = title)
  p <- ggplot(dat, aes(x=x, y=y)) +
    geom_point() +    
    stat_smooth(method = "lm") +
    facet_wrap(~serie) +
    ylab(y.name)
    xlab(x.name) +
    theme_ts() 
  return(p)
}
Number to text
}
\details{
Transform number to formatted text, 
the default is number rounded to 3 decimal places and 
separated by comma
}

