% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{outsample.mdd}
\alias{outsample.mdd}
\title{Previsao fora da amostra com modelo de defasagem distribuida}
\usage{
outsample.mdd(yh, yt, x, m = 3, p = 3, n, h = 12)
}
\arguments{
\item{yh}{serie a ser prevista \eqn{y_{t+h}^h}}

\item{yt}{(ts) preditor \eqn{y_{t-j+1}}}

\item{x}{(ts) preditor}

\item{m}{numero maximo de defasagens de x}

\item{p}{numero maximo de defasagens de yt}

\item{n}{numero de previsoes fora da amostra}

\item{h}{horizonte de previsao}
}
\value{
lista contendo \code{fcast} (ts) valores previstos; \code{model} 
(dyn) modelo estimado no fim da amostra
}
\description{
Realiza previsoes fora da amostra para yh com base no 
modelo de defasagem distribuida selecionado por BIC
}
