% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovecm.R
\name{outsample.vecm}
\alias{outsample.vecm}
\title{Previsao fora da amostra com VECM}
\usage{
outsample.vecm(y, max.p = 6, ic = c("SC", "HQ", "AIC", "FPE"),
  beta = NULL, include = c("const", "trend", "none", "both"),
  LRinclude = c("none", "const", "trend", "both"), seasonal = FALSE, h, k)
}
\arguments{
\item{y}{A multivariate time series}

\item{max.p}{Determines the highest lag order for lag length selection 
according to the choosen ic.}

\item{ic}{Information criterion to be used in model selection.}

\item{beta}{Imposed cointegrating value.}

\item{include}{Type of deterministic regressors to include}

\item{LRinclude}{Type of deterministic regressors to include in the 
long-term relationship. Can also be a matrix with exogeneous 
regressors (2OLS only)}

\item{seasonal}{If FALSE, restricts search to models without 
seasonal dummies.}

\item{h}{horizonte de previsao}

\item{k}{numero de previsoes fora da amostra}
}
\value{
uma lista com
fcast(ts): previsao pontual e intervalos de predicao usando arima 
outdate(num): vetor contendo periodo fora da amostra 
fit(autovecm): ultimo modelo VECM estimado
}
\description{
outsample.vecm e usada para gerar previsoes fora da amostra
usando modelo VECM estimado automaticamente com a funcao
autovecm
}

