% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovecm.R
\name{predict.autovecm}
\alias{predict.autovecm}
\title{Predict method for objects of class autovecm}
\usage{
\method{predict}{autovecm}(object, ..., n.ahead = 10, ci = 0.95)
}
\arguments{
\item{object}{An object of class 'autovecm'; generated by autovecm().}

\item{...}{Currently not used.}

\item{n.ahead}{An integer specifying the number of forecast steps.}

\item{ci}{The forecast confidence interval.}
}
\value{
A list with class attribute 'varprd' holding the following 
elements: 

  \item{\code{fcst}}{A list of matrices per variable containing 
  the forecasted values with lower and upper bounds 
  as well as the confidence interval.}
  \item{model}{The estimated VAR object for auto.var().}
}
\description{
Forecating a VECM object of class 'autovecm' with confidence bands
}

