% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{tab.comptest}
\alias{tab.comptest}
\title{Tabela com teste de previsao incorporada}
\usage{
tab.comptest(x, obs = "ipca12", ref = NULL, h)
}
\arguments{
\item{x}{matriz com previsoes nas colunas}

\item{obs}{nome da serie observada}

\item{ref}{nome da previsao de referencia}

\item{h}{Length ahead of the forecast.}
}
\value{
Uma matriz com a lambda estimado e o p valor do teste
de previsao incorporada
}
\description{
A partir de uma matrix x com as previsoes dos modelos e 
a series observada, retorna uma tabela com o teste de previsao
incorporada de Harvey et. al. 1998
}
