% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{tab.marques}
\alias{tab.marques}
\title{Tabela para as condicoes Marques et. al. (2003)}
\usage{
tab.marques(y, x)
}
\arguments{
\item{y}{um objeto ts com a inflacao observada}

\item{x}{um objeto mts com series dos nucleos a serem testados}
}
\value{
tabela com as estatisticas dos testes
}
\description{
Efetua o teste para as condicoes Marques et. al. (2003) para o nucleo e inflacao
 e fornece uma tabela com as estatisticas
}
