% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{tab.reqm}
\alias{tab.reqm}
\title{Tabela compara REQM dos modelos}
\usage{
tab.reqm(x, obs = "ipca", ref = "ipca.focus")
}
\arguments{
\item{x}{matriz com previsoes nas colunas}

\item{obs}{nome da serie observada}

\item{ref}{nome da previsao de referencia}
}
\value{
Uma matriz com REQM, EQMR e teste DM
}
\description{
A partir de uma matrix x com as previsoes dos modelos e 
a series observada, retorna uma tabela com REQM, REQM relativo
e teste dm
}
