% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{tab.stationary}
\alias{tab.stationary}
\title{Tabela para o teste de estacionaridade ADF e KPSS}
\usage{
tab.stationary(x, d = TRUE)
}
\arguments{
\item{x}{um objeto mts com series temporais multivariadas}

\item{d}{logico, TRUE (padrao) indica que teste deve ser 
efetuado tambem para series estacionarias}
}
\value{
tabela tabela com as estatisticas dos testes
}
\description{
Efetua o teste ADF e KPSS para series em nivel e 
diferenciadas e fornece uma tabela com as estatisticas
}
