% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-niemd.R
\name{tsplot}
\alias{tsplot}
\title{Plota series temporais com ggplot2}
\usage{
tsplot(x, y = NULL, escala = "free_y", facet = TRUE, name = NULL)
}
\arguments{
\item{x}{objeto ts ou mts com as series temporais}

\item{y}{(opicional) objeto ts ou mts com dimensao de x para ser
plotado junto com x no mesmo grafico}

\item{escala}{Are scales shared across all facets
("fixed"), or do they vary across 
rows ("free_x"), columns (the default, "free_y"), or both 
rows and columns ("free")}

\item{facet}{as series sao plotadas em graficos diferente (facet = TRUE, the default),
ou no mesmo grafico (facet = FALSE)}

\item{name}{optional name for ts univariate}
}
\value{
ggplot das series
}
\description{
Plota series temporais com ggplot2
}
